moved binding string extension to extensions folder
