+++ Sources/Playbook/Components/Text Area/PBTextArea.swift
@@ -174,14 +174,3 @@ struct PBTextArea_Previews: PreviewProvider {
       .background(Color.card)
   }
 }
-
-extension Binding where Value == String {
-  func max(_ limit: Int) -> Self {
-    if self.wrappedValue.count > limit {
-      DispatchQueue.main.async {
-        self.wrappedValue = String(self.wrappedValue.dropLast())
-      }
-    }
-    return self
-  }
-}
